requirements.txt
requests
pandas
mysql-connector-python
python-dotenv
jupyter
matplotlib
seaborn
wordcloud
scikit-learn
scipy
numpy
ipython
